Add liquabase to the project

+++ src/main/java/com/innova/security/services/UserDetailImpl.java
@@ -15,8 +15,7 @@ import java.util.stream.Collectors;
 public class UserDetailImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
 
-    private Long id;
-
+    private Integer id;
 
     private String username;
 
@@ -27,7 +26,7 @@ public class UserDetailImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailImpl(Long id,
+    public UserDetailImpl(Integer id,
                           String username, String email, String password,
                           Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
@@ -51,7 +50,7 @@ public class UserDetailImpl implements UserDetails {
         );
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
